% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reporting_data.R
\name{delphi_epidata_source}
\alias{delphi_epidata_source}
\title{Create a data source object for the Delphi Epidata API}
\usage{
delphi_epidata_source(
  target = c("covid", "flu", "rsv"),
  geo_types = c("state", "nation")
)
}
\arguments{
\item{target}{Character, disease target: "covid", "flu", or "rsv"}

\item{geo_types}{Character vector, geographic types to query (e.g., "state",
"nation")}
}
\value{
A source object of class "delphi_epidata_source"
}
\description{
Create a data source object for the Delphi Epidata API
}
