% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{dskellam_log}
\alias{dskellam_log}
\title{Skellam distribution log-probability mass function}
\usage{
dskellam_log(x, lambda1, lambda2)
}
\arguments{
\item{x}{Vector of integer observations (can be negative).}

\item{lambda1}{Rate parameter for the first Poisson (must be positive).}

\item{lambda2}{Rate parameter for the second Poisson (must be positive).}
}
\value{
Vector of log-probabilities.
}
\description{
Computes the log-PMF of the Skellam distribution, which is the distribution
of the difference of two independent Poisson random variables.
}
\keyword{internal}
