% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reporting_data.R
\name{epidata_source}
\alias{epidata_source}
\title{Create an epidata data source object}
\usage{
epidata_source(
  target = c("covid", "flu", "rsv"),
  geo_types = c("state", "nation")
)
}
\arguments{
\item{target}{Character, disease target: "covid", "flu", or "rsv"}

\item{geo_types}{Character vector, geographic types to query (e.g., "state",
"nation")}
}
\value{
A source object of class "epidata_source"
}
\description{
Create an epidata data source object
}
