% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reporting_data.R
\name{fetch_reporting_data.delphi_epidata_source}
\alias{fetch_reporting_data.delphi_epidata_source}
\title{Fetch reporting data from Delphi Epidata}
\usage{
\method{fetch_reporting_data}{delphi_epidata_source}(
  source,
  reference_dates,
  report_dates,
  locations = "*",
  ...
)
}
\arguments{
\item{source}{A delphi_epidata_source object}

\item{reference_dates}{Date vector or epirange of reference dates}

\item{report_dates}{Date vector or epirange of report dates}

\item{locations}{Character vector of locations ("*" for all)}

\item{...}{Additional arguments (unused)}
}
\value{
data.frame with reporting triangle data
}
\description{
Fetch reporting data from Delphi Epidata
}
\concept{data_fetching}
