% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_reporting_data_hubdata.R
\name{hub_data_source}
\alias{hub_data_source}
\title{Create a data source object for a forecast hub S3 bucket}
\usage{
hub_data_source(
  hub_name = "covid19-forecast-hub",
  target = "wk inc covid hosp"
)
}
\arguments{
\item{hub_name}{Character, the S3 bucket name (e.g.,
"covid19-forecast-hub").}

\item{target}{Character, the target name to filter to in the hub data
(e.g., "wk inc covid hosp" for NHSN COVID hospitalizations).}
}
\value{
A source object of class "hub_data_source"
}
\description{
Create a data source object for a forecast hub S3 bucket
}
\concept{data_sources}
