% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_methods.R
\name{sample_normal}
\alias{sample_normal}
\title{Sample from a normal distribution given predictions and variance}
\usage{
sample_normal(pred, uncertainty_params)
}
\arguments{
\item{pred}{Vector of predictions (means).}

\item{uncertainty_params}{Vector of variance parameters. If length 1,
the same variance is used for all predictions.}
}
\value{
Integer vector of samples of the same length as \code{pred}.
}
\description{
Generates samples from a normal distribution with the given predictions
as means and the estimated variance, rounded to integers for use with
count data. This is the sampling counterpart to \code{\link[=fit_normal]{fit_normal()}}.
}
\examples{
pred <- c(3.2, 4.6, 5.1)
variance <- 2.5
samples <- sample_normal(pred, uncertainty_params = variance)
samples
}
\seealso{
Other sample_distribution:
\code{\link[=sample_skellam]{sample_skellam()}}
}
\concept{probabilistic_methods}
\concept{sample_distribution}
