% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{fit_normal}
\alias{fit_normal}
\title{Fit a normal distribution with fixed mean to estimate variance}
\usage{
fit_normal(x, mu)
}
\arguments{
\item{x}{Vector of observed values.}

\item{mu}{Vector of expected values (fixed means).}
}
\value{
The maximum likelihood estimate of the variance (sigma^2).
}
\description{
Takes in a vector of observations and a vector of fixed means (predictions)
and returns the maximum likelihood estimate of the variance. With a fixed
mean, the MLE for variance is simply the mean squared error.
}
\examples{
obs <- c(4, 8, 10)
pred <- c(3.1, 7.2, 11)
variance <- fit_normal(obs, pred)
variance
}
\seealso{
Other estimate_observation_error: 
\code{\link{fit_skellam}()}
}
\concept{estimate_observation_error}
\concept{probabilistic_methods}
