% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcastNHSN-package.R
\docType{package}
\name{nowcastNHSN-package}
\alias{nowcastNHSN}
\alias{nowcastNHSN-package}
\title{nowcastNHSN: Nowcasting Tools for NHSN Data}
\description{
Provides utilities for nowcasting and working with National Healthcare Safety Network (NHSN) data, including date conversion utilities for epidemiological week formats.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
