% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_methods.R
\name{sample_skellam}
\alias{sample_skellam}
\title{Sample from a Skellam distribution given predictions and variance}
\usage{
sample_skellam(pred, uncertainty_params)
}
\arguments{
\item{pred}{Vector of predictions (means). Can be negative.}

\item{uncertainty_params}{Vector of variance parameters. Must satisfy
variance > |pred| for valid Poisson parameters. If length 1, the same
variance is used for all predictions.}
}
\value{
Integer vector of samples of the same length as \code{pred}.
}
\description{
Generates samples from a Skellam distribution (difference of two Poissons)
with the given predictions as means and the estimated variance.
This is the sampling counterpart to \code{\link[=fit_skellam]{fit_skellam()}}.

The Skellam distribution is parameterized by λ₁ and λ₂ where:
\itemize{
\item Mean = λ₁ - λ₂ = pred
\item Variance = λ₁ + λ₂ = uncertainty_params
}

The individual parameters are recovered as:
\itemize{
\item λ₁ = (pred + variance) / 2
\item λ₂ = (variance - pred) / 2
}
}
\examples{
pred <- c(3.2, -1.5, 0.8)
variance <- 10
samples <- sample_skellam(pred, uncertainty_params = variance)
samples
}
\seealso{
Other sample_distribution:
\code{\link{sample_normal}()}
}
\concept{probabilistic_methods}
\concept{sample_distribution}
